Final-project/requirements.txt
pytest-xdist==2.2.1
mimesis==4.1.3
requests==2.25.1
pytest==6.2.3
selenium==3.141.0
webdriver-manager==3.3.0
Flask==2.0.1
SQLAlchemy==1.4.17
allure-pytest==2.8.40
PyMySQL==1.0.2
